requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
ocrmypdf==15.4.0
python-multipart==0.0.6
pydantic==2.4.2
